Update product details and image handling
- Modify  function to include  inside the  parameter for image paths
- Ensure images fit 518x518 dimensions in
- Add styles for main image and thumbnails
- Update HTML structure in
- Implement image change and animation in
