+++ src/pages/prodoucts-view.html
@@ -134,11 +134,9 @@
     <footer>
       <iframe src="/src/components/footer.html" style="width: 100%; height: auto; border: none;" scrolling="no"></iframe>
     </footer>
-    <!-- Creating the script fot adding the way to change shoe after select one of them -->
-     <script src="/src/js/products-scripts/product-animation.js"></script>
-    <!--Script for generating the sizes-->
-    <script src="/src/js/products-scripts/sizes-generation.js"></script>
     <!--Script for move the page to the products section-->
     <script src="/src/js/products-scripts/send-from-index.js"></script>
+    <!--Script for generating the sizes-->
+    <script src="/src/js/products-scripts/sizes-generation.js"></script>
   </body>
 </html>
